+++ app/src/main/java/com/rodrigolmti/lunch/money/companion/features/transactions/ui/detail/ITransactionDetailUIModel.kt
@@ -8,7 +8,7 @@ import kotlinx.coroutines.flow.StateFlow
 internal interface ITransactionDetailUIModel {
     val viewState: StateFlow<TransactionDetailUiState>
 
-    fun getTransaction(id: Int)
+    fun getTransaction(id: Long)
 
     fun updateTransaction(
         update: UpdateTransactionView,
